`sappho-east`: split out recursive expr types into individual mods.
